Rename Balancer pool cache files
