+++ packages/asset-swapper/src/utils/market_operation_utils/pools_cache/index.ts
@@ -1,4 +1,4 @@
-export { BalancerPoolsCache } from './balancer_utils';
-export { BalancerV2PoolsCache } from './balancer_v2_utils';
-export { CreamPoolsCache } from './cream_utils';
+export { BalancerPoolsCache } from './balancer_pools_cache';
+export { BalancerV2PoolsCache } from './balancer_v2_pools_cache';
+export { CreamPoolsCache } from './cream_pools_cache';
 export { AbstractPoolsCache, PoolsCache } from './pools_cache';
