+++ packages/asset-swapper/src/utils/market_operation_utils/sampler_operations.ts
@@ -58,7 +58,7 @@ import {
 import { getGeistInfoForPair } from './geist_utils';
 import { getLiquidityProvidersForPair } from './liquidity_provider_utils';
 import { BalancerPoolsCache, BalancerV2PoolsCache, CreamPoolsCache, PoolsCache } from './pools_cache';
-import { BalancerV2SwapInfoCache } from './pools_cache/balancer_v2_utils_new';
+import { BalancerV2SwapInfoCache } from './pools_cache/balancer_v2_swap_info_cache';
 import { SamplerContractOperation } from './sampler_contract_operation';
 import { SamplerNoOperation } from './sampler_no_operation';
 import { SourceFilters } from './source_filters';
